==> election results/00_parse_cvk_data_round2_v1.ipynb <==
import bs4
import requests
import re
import json
import pandas as pd

from os import path
from tqdm import tqdm_notebook
from time import sleep

base_url = 'https://www.cvk.gov.ua/pls/vp2019/'
page = requests.get(path.join(base_url,'wp030pt001f01=719.html')).text
soup = bs4.BeautifulSoup(page,'html.parser')

district_links = []
for item in soup.find_all('tr', class_="tr"):
    district_links.append(item.find(class_ = 'a1').get('href'))
        
district_links[:10]

def parse_row(results):
    for row in results:
        for index,column in enumerate(row.find_all('td')[:2]):
            if index == 0:
                district_number = column.text.replace(u'\n', u'')
            elif index == 1:
                district_location = column.text.replace(u'\n', u'')
        records.append([district_general,district_number,district_location])

records = []

for district_link in tqdm_notebook(district_links):
    if district_link == 'wp045pt001f01=719.html':
        page = requests.get(path.join(base_url,district_link)).text
        soup = bs4.BeautifulSoup(page,'html5lib')
        fd_links = [link.get('href') for link in soup.find_all(class_="pure-table")[1].find_all(class_='a1')]

        for fd_link in fd_links:
            page = requests.get(path.join(base_url,fd_link)).text
            soup = bs4.BeautifulSoup(page,'html5lib')
            sleep(0.2)

            district_general = soup.find('h3').text
            results_raw = soup.find(class_='pure-table w100').find('tbody')
            parse_row(results_raw.find_all('tr'))
            
    else:
        page = requests.get(path.join(base_url,district_link)).text
        soup = bs4.BeautifulSoup(page,'html5lib')
        inner_link = soup.find(class_="fas fa-2x fa-warehouse").parent.get('href')

        page = requests.get(path.join(base_url,inner_link)).text
        soup = bs4.BeautifulSoup(page,'html5lib')
        sleep(0.2)

        district_general = soup.find(class_='pure-u-1 cntr').text.replace(u'\xa0', u'').replace(u'\n', u'')
        results_raw = soup.find(class_='pure-table pure-table-horizontal').find('tbody')
        parse_row(results_raw.find_all('tr')[1:])
    
colnames = ['district_general','district_number','district_location']
df_locations = pd.DataFrame.from_records(records,columns=colnames)
print(df_locations.shape)
df_locations.head()

base_url = 'https://www.cvk.gov.ua/pls/vp2019/'
page = requests.get(path.join(base_url,'wp335pt001f01=720.html')).text
soup = bs4.BeautifulSoup(page,'html5lib')
links = [link_raw.get('href') for link_raw in soup.find_all(class_='a1')]

districts = [int(link.split('=')[-1].replace('.html','')) for link in links]
links_and_districts = [i for i in zip(districts,links)]
links_and_districts = links_and_districts[1:]

    
districts_processed = [int(item.text.replace('\n',''))
                       for item in soup.find_all(class_='cntr b') 
                       if (item.get('style')!='border:1px solid #cbcbcb')&(item.text != 'ЗВО\n')]

print(len(districts_processed))
print(sum(districts_processed))

links_and_districts[:10]

colvalues = []
district_gen_numeric = []

for link in tqdm_notebook(links_and_districts):
    page = requests.get(path.join(base_url,link[1])).text
    soup = bs4.BeautifulSoup(page,'html5lib')
    results_raw = soup.find(class_='pure-table pure-table-horizontal w100')
    results_raw.find('thead').find_all('th')[0].text
    colnames = [colname.text.replace(u'\n', u'') for colname in results_raw.find('thead').find_all('th')]
    for row in results_raw.find_all('tr')[1:]:
        colvalues_row = []
        for colvalue in row.find_all('td'):
            colvalues_row.append(colvalue.text.replace(u'\n', u''))
        colvalues.append(colvalues_row)
        district_gen_numeric.append(link[0])
    sleep(0.2)

df_results = pd.DataFrame.from_records(colvalues,columns=colnames)
df_results['district_gen_numeric'] = pd.Series(district_gen_numeric)
print(sum(districts_processed)==df_results.shape[0])
print(df_results.shape)
df_results.head()

df_locations['district_gen_numeric'] = df_locations['district_general'].map(lambda x: int(x.split('№')[-1]) if len(x.split('№')[-1])<4 else 226)
df_locations['district_number'] = df_locations.loc[:,'district_number'].map(lambda x: x.replace(' ',''))
df_all = df_results.rename(columns={'№ ВД':'district_number'}).merge(df_locations,on=['district_number','district_gen_numeric'])
df_all = df_all[['district_number','district_gen_numeric', 'district_general','district_location'] + list(df_all.columns[1:-3])]
df_all['2.К-сть виборців, внесених до списку'] = pd.to_numeric(df_all['2.К-сть виборців, внесених до списку'])
print(df_all.shape)
df_all.head()

dtype_dic = {'district_number': str}
df_round1 = pd.read_csv('00_data/elections_round1_v1.csv',dtype = dtype_dic)
df_round1 = df_round1[['district_number','district_gen_numeric','2.К-сть виборців, внесених до списку']]
df_round1.head()

lost_distr = df_round1.merge(df_locations,on=['district_number','district_gen_numeric'],how='left')
lost_distr = lost_distr.loc[lost_distr['district_location'].isnull()]
print(lost_distr['2.К-сть виборців, внесених до списку'].sum())
lost_distr

df_all_compare = df_all.merge(df_round1, on=['district_number','district_gen_numeric'])
df_all_compare['reg_voters_change'] = df_all_compare['2.К-сть виборців, внесених до списку_x'] - df_all_compare['2.К-сть виборців, внесених до списку_y']
df_all_compare['reg_voters_change_abs'] = df_all_compare['reg_voters_change'].map(lambda x: abs(x))
print(df_all_compare['reg_voters_change'].sum())
print(df_all_compare['reg_voters_change_abs'].sum())
print(df_all_compare['reg_voters_change_abs'].sum()/29657746)
print(df_all_compare.shape)
df_all_compare.head()

df_all_compare.to_csv('00_data/elections_round2_v3.csv',index=False)
