==> election results/01_simulate_round2.ipynb <==
import pandas as pd
import numpy as np

dtype_dic= {'district_number': str}
df_round1 = pd.read_csv('00_data/elections_round1_v1.csv',dtype = dtype_dic)
print(df_round1.shape)
df_round1.head()

colnames_additional = ['2.К-сть виборців, внесених до списку',
                       '9.К-сть виборців, які взяли участь у голосуванні',
                       '10.К-сть бюлетенів, визнаних недійсними']
colnames = list(df_round1.columns[:4]) + colnames_additional
candidates = ['ЗеленськийВолодимир','ПорошенкоПетро']

df_round2 = df_round1[colnames+candidates].copy()
df_round2 = df_round2.rename(columns={candidate: candidate+'_round1' for candidate in candidates})
df_round2['participation_round1'] = (df_round2[colnames_additional[1]]-df_round2[colnames_additional[2]])/df_round2[colnames_additional[0]]
df_round2 = df_round2.drop([colnames_additional[1],colnames_additional[2]],axis=1)
df_round2['participation_delta'] = pd.Series(np.random.normal(scale=0.1,size=df_round2.shape[0]),name='participation_delta')
df_round2['participation_round2'] = df_round2['participation_round1'] + df_round2['participation_delta']
df_round2['participation_round2'] = df_round2['participation_round2'].map(lambda x: x if x>0 else 0.05)
df_round2['participation_round2'] = df_round2['participation_round2'].map(lambda x: x if x<0.95 else 0.95)
df_round2['participation_delta'] = df_round2['participation_round2'] - df_round2['participation_round1']
df_round2['voted_round2'] = (df_round2[colnames_additional[0]]*df_round2['participation_round2']).map(lambda x: round(x))
df_round2['zel_share'] = pd.Series(0.6+np.random.normal(loc=-0.05,scale=0.1,size=df_round2.shape[0]))
df_round2[candidates[0]] = (df_round2['voted_round2']*df_round2['zel_share']).map(lambda x: round(x))
df_round2[candidates[1]] = df_round2['voted_round2'] - df_round2[candidates[0]]

print(df_round2.shape)
df_round2.head()

df_round2.to_csv('00_data/elections_round2_sim_v7.csv')
