==> election results/02_json_manipulations_real_data.ipynb <==
import json
import os
import pandas as pd
import numpy as np

from tqdm import tqdm_notebook

files = os.listdir('tvk')
first = True

for file in tqdm_notebook(files):
    with open(os.path.join('tvk',file),'r', encoding="utf8") as cur_file:
        parsed_file = json.load(cur_file)
        if first:
            tvos_all = parsed_file
            first = False
        else:
            tvos_all['features'] = tvos_all['features'] + parsed_file['features']
            
len(tvos_all['features'])

dtype_dic = {'district_number': str}
df_results = pd.read_csv('00_data/elections_round2_all_v2.csv',dtype=dtype_dic)
df_results.shape

df_results.columns

df_results['participation_absolute_round1'] = df_results['2.К-сть виборців, внесених до списку']*df_results['participation_round1']
df_results['participation_absolute_round2'] = df_results['2.К-сть виборців, внесених до списку']*df_results['participation_round2']
df_results_tvo = df_results.groupby('district_gen_numeric')[['2.К-сть виборців, внесених до списку',
                                                             'participation_absolute_round1',
                                                             'participation_absolute_round2',
                                                             'ЗеленськийВолодимир',
                                                             'ПорошенкоПетро',
                                                             'ЗеленськийВолодимир_round2',
                                                             'ПорошенкоПетро_round2']].sum()
df_results_tvo['participation_diff'] = (df_results_tvo['participation_absolute_round2']-df_results_tvo['participation_absolute_round1'])/df_results_tvo['2.К-сть виборців, внесених до списку']
df_results_tvo['winner'] = (df_results_tvo['ЗеленськийВолодимир_round2'] - df_results_tvo['ПорошенкоПетро_round2']).map(lambda x: 'zel' if x>0 else 'por')
df_results_tvo['zel_growth'] = (df_results_tvo['ЗеленськийВолодимир_round2'] - df_results_tvo['ЗеленськийВолодимир'])/df_results_tvo['2.К-сть виборців, внесених до списку']
df_results_tvo['por_growth'] = (df_results_tvo['ПорошенкоПетро_round2'] - df_results_tvo['ПорошенкоПетро'])/df_results_tvo['2.К-сть виборців, внесених до списку']

print(df_results_tvo.shape)
df_results_tvo.head()

for feature in tqdm_notebook(tvos_all['features']):
    if int(feature['properties']['id']) in df_results_tvo.index:
        feature['properties']['participation_diff'] = df_results_tvo.loc[int(feature['properties']['id'])]['participation_diff']
        feature['properties']['zel_growth'] = df_results_tvo.loc[int(feature['properties']['id'])]['zel_growth']
        feature['properties']['por_growth'] = df_results_tvo.loc[int(feature['properties']['id'])]['por_growth']
        feature['properties']['winner'] = df_results_tvo.loc[int(feature['properties']['id'])]['winner']
    else:
        feature['properties']['participation_diff'] = 9999
        feature['properties']['zel_growth'] = 9999
        feature['properties']['por_growth'] = 9999
        feature['properties']['winner'] = 'no_vote'

len(tvos_all['features'])

with open('01_map/tvos_all_v6.json','w', encoding="utf8") as output:
    json.dump(tvos_all,output,ensure_ascii=False)

with open('newElectionData.json','r',encoding="utf8") as file:
    districts_all = json.load(file)

district = districts_all['objects']['-']['geometries'][1300]

for district in tqdm_notebook(districts_all['objects']['-']['geometries']):
    if district['properties']['d'] in list(df_results['district_number']):
        district['properties']['participation_diff'] = df_results.loc[df_results['district_number']==district['properties']['d']].reset_index()['participation_delta'][0]
    else:
        district['properties']['participation_diff'] = 9999

with open('newElectionData_v2.json','w',encoding="utf8") as file:
    json.dump(districts_all,file,ensure_ascii=False)
