==> election results/03_mobilization_real_data_v1.ipynb <==
import pandas as pd
import statsmodels.formula.api as smf
import numpy as np

from tqdm import tqdm_notebook
from scipy.stats import linregress
from scipy.optimize import minimize

candidates = ['ЗеленськийВолодимир','ПорошенкоПетро']

dtype_dic = {'district_number': str}
df_results = pd.read_csv('00_data/elections_round2_v3.csv',dtype=dtype_dic)
df_round1 = pd.read_csv('00_data/elections_round1_v1.csv',dtype = dtype_dic)
df_results = df_results.rename(columns={'ЗеленськийВолодимир':'ЗеленськийВолодимир_round2',
                                        'ПорошенкоПетро':'ПорошенкоПетро_round2',
                                        '2.К-сть виборців, внесених до списку_y':'2.К-сть виборців, внесених до списку'})

df_results['participation_round2'] = (df_results[candidates[0]+'_round2'] + df_results[candidates[1]+'_round2'])/df_results['2.К-сть виборців, внесених до списку']

'''
On several smaller voting districts (e.g., hospitals) inflow of new voters registrations
led to paricipation > 1 (because we take registration at round 1 as a baseline).
Therefore, is such cases minor adjustment is needed
'''
def adjust_participation(x):
    if x['participation_round2']>1:
        x['ЗеленськийВолодимир_round2'] = x['ЗеленськийВолодимир_round2']/x['participation_round2']
        x['ПорошенкоПетро_round2'] = x['ПорошенкоПетро_round2']/x['participation_round2']
        x['participation_round2'] = (x['ЗеленськийВолодимир_round2']+x['ПорошенкоПетро_round2'])/x['2.К-сть виборців, внесених до списку']
    return x['ЗеленськийВолодимир_round2'],x['ПорошенкоПетро_round2'],x['participation_round2']

df_results['ЗеленськийВолодимир_round2'], df_results['ПорошенкоПетро_round2'], df_results['participation_round2'] = zip(*df_results.apply(lambda x:adjust_participation(x),axis=1))
df_results.head()

print(df_results.shape)
print(df_round1.shape)
df_results.head()

df_round1['participation_round1'] = (df_round1['9.К-сть виборців, які взяли участь у голосуванні'] - 
                                     df_round1['10.К-сть бюлетенів, визнаних недійсними'])/df_round1['2.К-сть виборців, внесених до списку']
df_round1.head()

base_cols = ['district_number', 'district_gen_numeric']

other_cols = ['participation_round1']

main_cands = ['БойкоЮрій',
             'ВілкулОлександр',
             'ГриценкоАнатолій',
             'ЛяшкоОлег',
             'СмешкоІгор',
             'ТимошенкоЮлія',
             'other_cands']

other_cands = ['БалашовГеннадій',
               'БезсмертнийРоман', 
               'БогомолецьОльга', 
               'БогословськаІнна',
               'БондарВіктор',
               'ВащенкоОлександр',
               'ГаберМикола',
               'ДанилюкОлександр',
               'Дерев’янкоЮрій',
               'ЖуравльовВасиль',
               'КаплінСергій',
               'КармазінЮрій',
               'КиваІлля',
               'КорнацькийАркадій',
               'КошулинськийРуслан',
               'КривенкоВіктор',
               'КупрійВіталій',
               'ЛитвиненкоЮлія',
               'МорозОлександр',
               'НаливайченкоВалентин',
               'НасіровРоман',
               'НовакАндрій',
               'НосенкоСергій',
               'ПетровВолодимир',
               'РиговановРуслан',
               'СкоцикВіталій',
               'СоловйовОлександр',
               'ТарутаСергій',
               'ТимошенкоЮрій',
               'ШевченкоІгор',
               'ШевченкоОлександр']

df_round1['other_cands'] = df_round1[other_cands].sum(axis=1)
df_round1['participation_round1'] = (df_round1['9.К-сть виборців, які взяли участь у голосуванні'] - df_round1['10.К-сть бюлетенів, визнаних недійсними'])/df_round1['2.К-сть виборців, внесених до списку']
df_candidates1 = df_round1[base_cols+other_cols+candidates+main_cands].copy()
df_candidates1.loc[:,'pool_round1'] = df_candidates1.loc[:,main_cands].sum(axis=1)
print(df_candidates1.shape)
df_candidates1.head()

df_all = df_results.merge(df_candidates1,on=base_cols)
print(df_all.shape)
print(df_all.columns)
df_all.head()

#stage 1: find out, who stayed home or mobilized
df_all['stayed_home'] = df_all['participation_round1']*df_all['2.К-сть виборців, внесених до списку']-df_all['participation_round2']*df_all['2.К-сть виборців, внесених до списку']
df_all['mobilized'] = df_all['stayed_home'].map(lambda x: x*(-1) if x<0 else 0)
df_all['stayed_home'] = df_all['stayed_home'].map(lambda x: x if x>0 else 0)

df_all['voted_round1'] = df_all['2.К-сть виборців, внесених до списку']*df_all['participation_round1']
df_all['voted_round2'] = df_all['2.К-сть виборців, внесених до списку']*df_all['participation_round2']

for candidate in candidates:
    df_all[candidate+'_stayed_home'] = df_all.apply(lambda x: x[candidate]-x[candidate+'_round2'] if ((x[candidate]-x[candidate+'_round2'])>0)&(x['stayed_home']>0) else 0,axis=1)

df_all['stayed_home_others'] = df_all['stayed_home']-df_all[candidates[0]+'_stayed_home']-df_all[candidates[1]+'_stayed_home']
    
for candidate in main_cands:
    share = df_all[candidate]/df_all['pool_round1']
    df_all[candidate+'_stayed_home'] = share*df_all['stayed_home_others']

def mobilization(x):
    share_zel = x[candidates[0]+'_round2']/x['voted_round2']
    if (x[candidates[0]]>x[candidates[0]+'_round2']):
        return (0,x['mobilized'])
    elif (x[candidates[1]]>x[candidates[1]+'_round2']):
        return(x['mobilized'],0)
    elif (x[candidates[0]]+x['mobilized']*share_zel)>x[candidates[0]+'_round2']:
        part_mobilization = x[candidates[0]+'_round2']-x[candidates[0]]
        return(part_mobilization,x['mobilized']-part_mobilization)
    elif (x[candidates[1]]+x['mobilized']*(1-share_zel))>+x[candidates[1]+'_round2']:
        part_mobilization = x[candidates[1]+'_round2']-x[candidates[1]]
        return(x['mobilized']-part_mobilization,part_mobilization)
    else:
        return(x['mobilized']*share_zel,x['mobilized']*(1-share_zel))

df_all[candidates[0]+'_mobilized'],df_all[candidates[1]+'_mobilized'] = zip(*df_all.apply(lambda x: mobilization(x), axis=1))
    
print(f'Shape after stage 1: {df_all.shape}')

#stage 2: find out, where votes outflowed to the competitor
for candidate in candidates:
    share = df_all[candidate]/df_all['voted_round1']
    df_all[candidate+'_outflow'] = df_all[candidate]-df_all[candidate+'_round2']-df_all[candidate+'_stayed_home']
    df_all[candidate+'_outflow'] = df_all[candidate+'_outflow'].map(lambda x: x if x>0 else 0)
    
df_all['ЗеленськийВолодимир_inflow'] = df_all['ПорошенкоПетро_outflow']
df_all['ПорошенкоПетро_inflow'] = df_all['ЗеленськийВолодимир_outflow']
print(f'Shape after stage 2: {df_all.shape}')

#stage 3: find out, how the votes from the other candidates inflowed to the candidate
for candidate in candidates:
    df_all[candidate+'_inflow_others'] = df_all[candidate+'_round2']-df_all[candidate]-df_all[candidate+'_inflow']-df_all[candidate+'_mobilized']+df_all[candidate+'_outflow']+df_all[candidate+'_stayed_home']
    for candidate_lost in main_cands:
        share = df_all[candidate_lost]/df_all['pool_round1']
        df_all[candidate+'_inflow_'+candidate_lost] = share*df_all[candidate+'_inflow_others']
print(f'Shape after stage 3: {df_all.shape}')

df_all.head()

'''
Basic checks
'''
#voters quantity check
print(f"Voters check round 1 error: {df_all['voted_round1'].sum() - sum([df_all[i].sum() for i in ['ЗеленськийВолодимир', 'ПорошенкоПетро','pool_round1']])}")
print(f"Voters check round 2 error: {df_all['voted_round2'].sum() - sum([df_all[candidate+'_round2'].sum() for candidate in candidates])}")

#mobilization check
print(f"Mobilization between rounds error: {df_all['voted_round2'].sum() - (df_all['voted_round1'].sum() + df_all['mobilized'].sum() - df_all['stayed_home'].sum())}")
if df_all['mobilized'].sum() == df_all[candidates[0]+'_mobilized'].sum() + df_all[candidates[1]+'_mobilized'].sum():
    print(f'Mobilization distribution check is successful')
else:
    print(f"Mobilization distribution check error: {df_all['mobilized'].sum() - (df_all[candidates[0]+'_mobilized'].sum() + df_all[candidates[1]+'_mobilized'].sum())}")

for candidate in candidates:
    #negative values check
    if len(df_all[candidate].loc[df_all[candidate]<-1])>0:
        print(f'Round1 data on {candidate} has negtative values')
        break
    elif len(df_all[candidate+'_inflow'].loc[df_all[candidate+'_inflow']<-1])>0:
        print(f'Inflow data on {candidate} has negtative values')
        break
    elif len(df_all[candidate+'_inflow_others'].loc[df_all[candidate+'_inflow_others']<-1])>0:
        print(f'Inflow (others) data on {candidate} has negtative values')
        break
    elif len(df_all[candidate+'_stayed_home'].loc[df_all[candidate+'_stayed_home']<-1])>0:
        print(f'Stayed_home data on {candidate} has negtative values')
        break
    elif len(df_all[candidate+'_outflow'].loc[df_all[candidate+'_outflow']<-1])>0:
        print(f'Outflow data on {candidate} has negtative values')
        break

    print(f'Negative values check for {candidate} is successful')
    
    #decomposition check
    round1 = df_all[candidate].sum()
    round2 = df_all[candidate+'_round2'].sum()

    inflow = df_all[candidate+'_inflow'].sum()
    inflow_others = df_all[candidate+'_inflow_others'].sum()
    mobilized = df_all[candidate+'_mobilized'].sum()

    home = df_all[candidate+'_stayed_home'].sum()
    outflow = df_all[candidate+'_outflow'].sum()

    if round2 == round1 + inflow + inflow_others + mobilized - home - outflow:
        print(f'Voters decomposition for candidate {candidate} is successful')
    else:
        print(f'Voters decomposition for candidate {candidate}: {round2-(round1 + inflow + inflow_others + mobilized - home - outflow)} error')

    #inflows check
    check_inflows = 0
    for candidate_lost in main_cands:
        check_inflows += df_all[candidate+'_inflow_'+candidate_lost].sum()

    if inflow_others == check_inflows:
        print(f'Inflows test for candidate {candidate} is successful')
    else:
        print(f'Inflows test for candidate {candidate}: {inflow_others-check_inflows} error')

df_all.to_csv('00_data/elections_round2_all_v2.csv',index=False)

for candidate in main_cands:
    df_all_copy[candidate+'_inflow'] = df_all_copy[candidate] - df_all_copy[candidate+'_stayed_home']

string = ' + '
cands_inflow = [cand+'_inflow' for cand in main_cands]

results = smf.ols('ЗеленськийВолодимир_inflow_others ~ ' + string.join(cands_inflow), data=df_all_copy).fit()
print(results.summary())

results2 = smf.ols('ПорошенкоПетро_inflow_others ~ ' + string.join(cands_inflow), data=df_all_copy).fit()
print(results2.summary())

bounds = [(0,1) for cand in main_cands]
bounds

X = df_all_copy[cands_inflow]
print(X.shape)

# Define the Model
model = lambda b, X: b[0] * X.iloc[:,0] + b[1] * X.iloc[:,1] + b[2] * X.iloc[:,2] + b[3] * X.iloc[:,3] + b[4] * X.iloc[:,4] + b[5] * X.iloc[:,5] + b[6] * X.iloc[:,6]

# The objective Function to minimize (least-squares regression)
obj = lambda b, Y, X: np.sum(np.abs(Y-model(b, X))**2)

# Bounds: b[0], b[1], b[2] >= 0
bnds = [(0, 1) for i in cands_inflow]

# Initial guess for b[1], b[2], b[3]:
xinit = np.zeros(len(cands_inflow))

for candidate in candidates:

    Y = df_all_copy[candidate+'_inflow_others']
    res = minimize(obj, args=(Y, X), x0=xinit, bounds=bnds)
    
    print(candidate + ": " + string.join([main_cands[i] + " " + str(res.x[i]) for i in range(len(main_cands))]))

X.iloc[:,2]
