==> election results/04_viz_real_data.ipynb <==
import pandas as pd
import plotly.plotly as py
import plotly.graph_objs as go
import numpy as np

from plotly.offline import download_plotlyjs, init_notebook_mode, plot, iplot
from tqdm import tqdm_notebook

init_notebook_mode(connected=True)

df_all = pd.read_csv('00_data/elections_round2_all_v2.csv')
print(df_all.shape)
df_all.head()

#credentials needed
py.sign_in('okuchma', '1X4AOENqX8gafp4qNWcS')

candidates = ['ЗеленськийВолодимир','ПорошенкоПетро']

base_cols = ['district_number', 'district_gen_numeric']

main_cands = ['БойкоЮрій',
             'ВілкулОлександр',
             'ГриценкоАнатолій',
             'ЛяшкоОлег',
             'СмешкоІгор',
             'ТимошенкоЮлія',
             'other_cands']

no_vote = 'не голосували'
indices = [i for i in range(len(candidates+main_cands+[no_vote]+candidates+[no_vote]))]
indices = [i for i in zip([outflow for outflow in (candidates+main_cands+[no_vote])]+
                          [inflow for inflow in (candidates+[no_vote])],indices)]

indices

proper_names = {'ЗеленськийВолодимир':'Володимир Зеленський',
                'ПорошенкоПетро':'Петро Порошенко',
                'БойкоЮрій':'Юрій Бойко',
                'ВілкулОлександр':'Олександр Вілкул',
                'ГриценкоАнатолій':'Анатолій Гриценко',
                'ЛяшкоОлег':'Олег Ляшко',
                'СмешкоІгор':'Ігор Смешко',
                'ТимошенкоЮлія':'Юлія Тимошенко',
                'other_cands':'інші кандидати',
                'не голосували':'не голосували'}

labels = [proper_names[i[0]] for i in indices]
print(labels)

outflows = []
for index in indices[:10]:
    outflows += [index for i in range(3)]
    
inflows = [] 
for index in range(len(indices[:10])):
    inflows += indices[-3:]
    
colors_dict = {'Володимир Зеленський':'#3ec258',
               'Петро Порошенко':'#c2443e',
               'не голосували':'#666666'}

colors = [colors_dict[label] if label in colors_dict else '#ffdb83' for label in labels]

print(len(outflows),len(inflows))
source = [outflow[1] for outflow in outflows]
target = [inflow[1] for inflow in inflows]

for candidate in candidates:
    df_all[candidate+'_inflow_'+candidate] = df_all[candidate] - df_all[candidate+'_stayed_home'] - df_all[candidate+'_outflow']

df_all[candidates[0]+'_inflow_'+candidates[1]] = df_all[candidates[0]+'_inflow']
df_all[candidates[1]+'_inflow_'+candidates[0]] = df_all[candidates[1]+'_inflow']

df_all.columns

def count_values(outflow,inflow):
    outflow_name = outflow[0].replace('_outflow','')
    inflow_name = inflow[0].replace('_inflow','')
    no_vote1_outflow = df_all['2.К-сть виборців, внесених до списку'].sum()-df_all['voted_round1'].sum()-df_all['mobilized'].sum()
    
    print(f'From candidate {outflow_name} to candidate {inflow_name}')
    print(inflow_name+'_inflow_'+outflow_name)
    
    if (outflow_name==no_vote)|(inflow_name==no_vote):
        if (outflow_name==no_vote)&(inflow_name==no_vote):
            print(no_vote1_outflow)
            values.append(no_vote1_outflow)
        elif outflow_name==no_vote:
            print(df_all[inflow_name+'_mobilized'].sum())
            values.append(df_all[inflow_name+'_mobilized'].sum())
        else:
            print(df_all[outflow_name+'_stayed_home'].sum())
            values.append(df_all[outflow_name+'_stayed_home'].sum())
    else:
        print(df_all[inflow_name+'_inflow_'+outflow_name].sum())
        values.append(df_all[inflow_name+'_inflow_'+outflow_name].sum())
    
values = []
for index,value in enumerate(outflows):
    count_values(outflows[index],inflows[index])
sum(values)

values_share = [value/sum(values) for value in values]
values_share

np_source = np.array(source)
np_values = np.array(values)
np_maxes = np.array([])

for triplet in range(np_source.max()+1):
    np_maxes = np.append(np_maxes,np_values[np.where(np_source==triplet)].max())

color_links = np.array(['rgba(204,204,204,0.3)' for i in range(len(np_source))])
color_links[np.where(np.isin(np_values,np_maxes))[0]]='rgba(204,204,204,1)'

data = {'type':'sankey',
       'node':{'label':labels,
               'color':colors,
               'pad':10,
               'line':{'width':0},
               'hovertemplate': "<extra>%{value:.1%}</extra> від усіх зареєстрованих виборців"
              },
       'link':{'source':source,
              'target':target,
              'value':values_share,
              'color':color_links,
              'hovertemplate': "<extra>%{value:.2%}</extra> від усіх зареєстрованих виборців"}
       }

fig = {'data':[data]}
py.iplot(fig, validate=False)
#iplot(fig, validate=False)

['ЗеленськийВолодимир', 'ПорошенкоПетро',
'ЗеленськийВолодимир_stayed_home', 'ПорошенкоПетро_stayed_home',
'ЗеленськийВолодимир_outflow', 'ПорошенкоПетро_outflow',
'ЗеленськийВолодимир_inflow', 'ПорошенкоПетро_inflow',
'ЗеленськийВолодимир_inflow_others','ПорошенкоПетро_inflow_others',
'ЗеленськийВолодимир_mobilized', 'ПорошенкоПетро_mobilized']

candidates = ['Зеленський','Порошенко']
stayed_home = [df_all['ЗеленськийВолодимир_stayed_home'].sum()*(-1),df_all['ПорошенкоПетро_stayed_home'].sum()*(-1)]
outflow = [df_all['ЗеленськийВолодимир_outflow'].sum()*(-1),df_all['ПорошенкоПетро_outflow'].sum()*(-1)]
base = [df_all['ЗеленськийВолодимир'].sum(),df_all['ПорошенкоПетро'].sum()]
mobilization = [df_all['ЗеленськийВолодимир_mobilized'].sum(),df_all['ПорошенкоПетро_mobilized'].sum()]
inflow_competitor = [df_all['ЗеленськийВолодимир_inflow'].sum(),df_all['ПорошенкоПетро_inflow'].sum()]
inflow_others = [df_all['ЗеленськийВолодимир_inflow_others'].sum(),df_all['ПорошенкоПетро_inflow_others'].sum()]

labels_bar = ['0','2 млн','4 млн','6 млн','8 млн','10 млн']
tickvals = [0,2*10**6,4*10**6,6*10**6,8*10**6,10*10**6]

trace1 = go.Bar(
    y=candidates,
    x=stayed_home,
    orientation = 'h',
    marker = dict(
        color = ['rgba(62,194,88,0)','rgba(194,68,62,0)']
    )
)

trace2 = go.Bar(
    y=candidates,
    x=outflow,
    orientation = 'h',
    marker = dict(
        color = ['rgba(62,194,88,0)','rgba(194,68,62,0)']
    )
)

trace3 = go.Bar(
    y=candidates,
    x=base,
    orientation = 'h',
    marker = dict(
        color = ['rgba(62,194,88,1)','rgba(194,68,62,1)']
    )
)

trace4 = go.Bar(
    y=candidates,
    x=mobilization,
    orientation = 'h',
    marker = dict(
        color = ['rgba(62,194,88,0.8)','rgba(194,68,62,0.8)']
    )
)

trace5 = go.Bar(
    y=candidates,
    x=inflow_competitor,
    orientation = 'h',
    marker = dict(
        color = ['rgba(62,194,88,0.6)','rgba(194,68,62,0.6)']
    )
)

trace6 = go.Bar(
    y=candidates,
    x=inflow_others,
    orientation = 'h',
    marker = dict(
        color = ['rgba(62,194,88,0.6)','rgba(194,68,62,0.6)']
    )
)

data = [trace1, trace2, trace3, trace4, trace5, trace6]
layout = go.Layout(
    barmode='stack',
    showlegend=False,
    xaxis=go.layout.XAxis(
        ticktext=labels_bar,
        tickvals=tickvals
    )
)

fig = go.Figure(data=data, layout=layout)
#py.iplot(fig)
iplot(fig)
